# spec/openapi.yaml
openapi: 3.0.3
info:
  title: MealForward B2B AX Platform API
  description: >
    LLM Guardrail AX 기반 세무 리스크 없는 B2B 식대 이월 및 복지 정산 플랫폼 API 명세서.
    비과세 식대 포인트의 이월/사용/정산 흐름과, 현금성 자산 결제를 자동 차단하는
    AI 세무 가드레일(Guardrail AX) 엔드포인트를 포함한다.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Dev Server
...
paths:
...
  /user/meal-balance:
    get:
      tags: [User]
      summary: 당월 미사용 식대 잔액 및 연간 마감 D-Day 조회
      operationId: getMealBalance
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
          example: 1004
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealBalance'
  /user/points:
    get:
      tags: [User]
      summary: 보유 포인트 조회 (일반 복지 포인트 / 식대 이월 포인트 분리 표시)
      operationId: getUserPoints
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
          example: 1004
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                    example: 1004
                  generalPoint:
                    type: number
                    example: 50000.00
                  flexiMealPoint:
                    type: number
                    example: 32000.00
                  flexiMealExpiredAt:
                    type: string
                    example: "2027-09-30T23:59:59Z"
  /user/rollover/apply:
    post:
      tags: [User]
      summary: 차월 복지 포인트로 식대 이월 신청
      operationId: applyRollover
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, yearMonth]
              properties:
                userId:
                  type: integer
                  example: 1004
                yearMonth:
                  type: string
                  example: "202609"
      responses:
        '200':
          description: 이월 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  rolledAmount:
                    type: number
                    example: 32000.00
                  targetPointType:
                    type: string
                    example: "FLEXI_MEAL"
                  expiredAt:
                    type: string
                    example: "2027-09-30T23:59:59Z"
        '409':
          description: 이미 이월 완료된 당월 식대
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
  /user/purchases:
    get:
      tags: [User]
      summary: 구매 내역 조회 (사용하기/환불하기 대상)
      operationId: getUserPurchases
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
          example: 1004
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 501
                    itemName:
                      type: string
                      example: "신세계 상품권 5만원권"
                    generalAmount:
                      type: number
                      example: 20000.00
                    flexiAmount:
                      type: number
                      example: 30000.00
                    totalAmount:
                      type: number
                      example: 50000.00
                    approvalStatus:
                      type: string
                      example: "APPROVED"
                    redemptionStatus:
                      type: string
                      enum: [UNUSED, USED, REFUNDED]
                      example: "UNUSED"
                    voucherCode:
                      type: string
                      nullable: true
                      description: "사용하기 이후에만 노출됩니다."
                      example: null
                    createdAt:
                      type: string
                      example: "2026-09-10T04:20:00Z"
  /user/purchases/{id}/use:
    post:
      tags: [User]
      summary: 구매 건 사용 처리 (바우처/숫자 코드 발급 확인)
      operationId: useUserPurchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 501
      responses:
        '200':
          description: 사용 완료, 바우처 코드 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  voucherCode:
                    type: string
                    example: "4821-9012-3345-7710"
        '409':
          description: 이미 사용했거나 환불된 구매
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
  /user/purchases/{id}/refund:
    post:
      tags: [User]
      summary: 구매 건 환불 (결제에 사용한 포인트 종류로만 환불, 종류 간 교차 환불 불가)
      operationId: refundUserPurchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 501
      responses:
        '200':
          description: 환불 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  refundedGeneralAmount:
                    type: number
                    example: 20000.00
                  refundedFlexiAmount:
                    type: number
                    example: 30000.00
        '409':
          description: 이미 사용했거나 환불된 구매
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictError'
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: header
      name: x-admin-id
      description: 로그인한 관리자의 id. /auth/login(role=admin) 응답의 id 값을 그대로 전달한다.
  schemas:
...
    MealBalance:
      type: object
      properties:
        userId:
          type: integer
          example: 1004
        yearMonth:
          type: string
          example: "202609"
        remainingAmount:
          type: number
          example: 32000.00
        isRolledOver:
          type: boolean
          example: false
        expirationDate:
          type: string
          example: "2026-09-30T23:59:59Z"
...
    ConflictError:
      type: object
      properties:
        errorCode:
          type: string
          example: "DUPLICATE_ROLLOVER_REQUEST"
        message:
          type: string
          example: "이미 당월 이월 처리가 완료되었습니다."